fix: move SchallCircuitGroup integration to data-final-fixes phase
