+++ changelog.txt
@@ -1,4 +1,9 @@
 ---------------------------------------------------------------------------------------------------
+Version: 1.0.1
+Date: ????
+  Bugfixes:
+    - Move SchallCircuitGroup integration to data-final-fixes phase to ensure it applies correctly.
+---------------------------------------------------------------------------------------------------
 Version: 1.0.0
 Date: 2025-09-17
   Changes:
