+++ info.json
@@ -4,7 +4,7 @@
   "description": "Opinionated tweaks to Cybersyn Combinator.",
   "author": "QuingKhaos",
   "homepage": "",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "factorio_version": "2.0",
   "dependencies": [
     "base >= 2.0.0",
